requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
pytz==2023.3